Created CreatePostViewModel.kt and moved AuthActivity.kt to auth pacakage
