+++ app/src/main/AndroidManifest.xml
@@ -12,8 +12,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SocialNetworkClone"
         tools:ignore="AllowBackup">
-        <activity android:name=".ui.main.MainActivity"></activity>
-        <activity android:name=".AuthActivity">
+        <activity android:name=".ui.main.MainActivity" />
+        <activity android:name=".ui.auth.AuthActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
