+++ app/src/main/res/layout/activity_auth.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AuthActivity">
+    tools:context=".ui.auth.AuthActivity">
 
     <FrameLayout
         android:id="@+id/fragmentContainer"
